refactor: 파일 이동
- app.js에 있는 utils 함수 utils.js로 이동
