+++ src/js/__tests__/utils.test.js
@@ -3,7 +3,7 @@ import {
   isDivisibleBy,
   createRandomNumber,
   createRandomNumberList,
-} from '../app.js';
+} from '../utils.js';
 
 test('구입 금액은 양의 정수이다. 실패 케이스', () => {
   const payment = -1;
